pid_file: Do not attempt to free the read-only default string
If pid_file is not set through the config file the default value
remain in it and it resides in a read-only section. In case of a
CTRL+C it will be freed leading to a detected memory corruption.

Modify the code to strdup the default value in case pid_file is
still NULL. It means that after init pid_file is never NULL but
that seems to be the case right now as well.