pcap: Add initial version of sctp support
Add initial support for SCTP and parsing SCTP data chunks. It has
been "verified" with a single SCTP packet containing exactly one
chunk. The code does not support fragmentation of data chunks, the
input validation might not be enough. I didn't understand the
frag_offset calculation (as once we have the SCTP header all IP
offsets are already applied?).

In contrast to TCP and UDP a single frame can contain multiple
chunks and these are dispatched one after another. The counters
are for sctp frames. In general "full_packet" as an option does
not make sense and the hdrlen has not been set. It might require
more architecture changes to move forward. The code also does not
attempt to identify retransmits of a data chunk.

Extend the msg structure to transport the ppid which will allow
to parse M2UA in the next step.

Manaually tested with:

$ ./src/captagent -f ./conf/captagent.xml -n -D ../../single_isup_msg.pcap  -c