transport_hep: Store hep_connection inside the event loop
Using tls with uv_key_t requires a version of libuv not available
on Debian Jessie/Wheezy. Instead of storing the hep_connection as
TLS variable, store it inside the uv_loop_t as the data pointer.

All handles will (in-)directly refer to the eventloop and we can
access the hep_connection. The change is untested and might not
work (but in principle it should).